replace rewards plugin
